backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psutil==5.9.6
httpx==0.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.12.1
structlog==23.2.0
redis==5.0.1
apscheduler==3.10.4
google-genai==0.2.2
numpy==1.26.3